aws_batch/image/requirements.txt
boto3==1.28.85
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.1
selenium==4.15.2
tqdm==4.65.0
requests==2.31.0
beautifulsoup4==4.12.2
undetected-chromedriver==3.5.4
PyVirtualDisplay==3.0
chromedriver-autoinstaller==0.6.4